requirements.txt
discord.py>=2.0.0
langcodes
pycountry
aiomysql
requests
aiohttp